2D GUI Stuff, linear layouts with resize divider in between

+++ FFT.cpp
@@ -4,17 +4,17 @@
 
 using namespace omen;
 
-FFT::FFT(unsigned int N) : N(N), reversed(0), T(0), pi2(2 * M_PI) {
+FFT::FFT(unsigned int N) : N(N), reversed(0), T(0), pi2(static_cast<float>(2 * M_PI)) {
 	c[0] = c[1] = 0;
 
-	log_2_N = log(N) / log(2);
+	log_2_N = static_cast<int>(log(N) / log(2));
 
 	reversed = new unsigned int[N];		// prep bit reversals
-	for (int i = 0; i < N; i++) reversed[i] = reverse(i);
+	for (unsigned int i = 0; i < N; i++) reversed[i] = reverse(i);
 
 	int pow2 = 1;
 	T = new Complex*[log_2_N];		// prep T
-	for (int i = 0; i < log_2_N; i++) {
+	for (unsigned int i = 0; i < log_2_N; i++) {
 		T[i] = new Complex[pow2];
 		for (int j = 0; j < pow2; j++) T[i][j] = t(j, pow2 * 2);
 		pow2 *= 2;
@@ -29,7 +29,7 @@ FFT::~FFT() {
 	if (c[0]) delete[] c[0];
 	if (c[1]) delete[] c[1];
 	if (T) {
-		for (int i = 0; i < log_2_N; i++) if (T[i]) delete[] T[i];
+		for (unsigned int i = 0; i < log_2_N; i++) if (T[i]) delete[] T[i];
 		delete[] T;
 	}
 	if (reversed) delete[] reversed;
@@ -37,7 +37,7 @@ FFT::~FFT() {
 
 unsigned int FFT::reverse(unsigned int i) {
 	unsigned int res = 0;
-	for (int j = 0; j < log_2_N; j++) {
+	for (unsigned int j = 0; j < log_2_N; j++) {
 		res = (res << 1) + (i & 1);
 		i >>= 1;
 	}
@@ -45,17 +45,17 @@ unsigned int FFT::reverse(unsigned int i) {
 }
 
 Complex FFT::t(unsigned int x, unsigned int N) {
-	return Complex(cos(pi2 * x / N), sin(pi2 * x / N));
+	return Complex(static_cast<float>(cos(pi2 * x / N)), static_cast<float>(sin(pi2 * x / N)));
 }
 
 void FFT::fft(Complex* input, Complex* output, int stride, int offset) {
-	for (int i = 0; i < N; i++) c[which][i] = input[reversed[i] * stride + offset];
+	for (unsigned int i = 0; i < N; i++) c[which][i] = input[reversed[i] * stride + offset];
 
 	int loops = N >> 1;
 	int size = 1 << 1;
 	int size_over_2 = 1;
 	int w_ = 0;
-	for (int i = 1; i <= log_2_N; i++) {
+	for (unsigned int i = 1; i <= log_2_N; i++) {
 		which ^= 1;
 		for (int j = 0; j < loops; j++) {
 			for (int k = 0; k < size_over_2; k++) {
@@ -74,5 +74,5 @@ void FFT::fft(Complex* input, Complex* output, int stride, int offset) {
 		w_++;
 	}
 
-	for (int i = 0; i < N; i++) output[i * stride + offset] = c[which][i];
+	for (unsigned int i = 0; i < N; i++) output[i * stride + offset] = c[which][i];
 }
